+++ src/core/app.cpp
@@ -6,7 +6,9 @@
 #include <sstream>
 
 namespace fluxgl {
-    App::App(int width, int height, const char *title) : m_window(width, height, title) { }
+    App::App(int width, int height, const char *title) : m_window(width, height, title) {
+        m_window.bindApp(*this);
+    }
 
     void App::run() {
         try {
@@ -18,9 +20,12 @@ namespace fluxgl {
                 float deltaTime = currentTime - lastTime;
                 lastTime = currentTime;
 
+                m_inputManager.beginFrame();
                 m_window.pollEvents();
+
                 onUpdate(deltaTime);
                 onRender();
+
                 m_window.swapBuffers();
             }
 
@@ -37,5 +42,21 @@ namespace fluxgl {
             FLUXGL_LOG_ERROR("Unknown error occurred");
             exit(1);
         }
+    }
+
+    fluxgl::InputManager& App::getInput() { 
+        return m_inputManager; 
+    }
+
+    const fluxgl::InputManager& App::getInput() const { 
+        return m_inputManager;
+    }
+
+    fluxgl::Window& App::getWindow() { 
+        return m_window; 
+    }
+
+    const fluxgl::Window& App::getWindow() const { 
+        return m_window; 
     }
 }